% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Acl}
\alias{Acl}
\title{Google+ Domains API Objects 
Builds on top of the Google+ platform for Google Apps Domains.}
\usage{
Acl(description = NULL, domainRestricted = NULL, items = NULL)
}
\arguments{
\item{description}{Description of the access granted, suitable for display}

\item{domainRestricted}{Whether access is restricted to the domain}

\item{items}{The list of access entries}
}
\value{
Acl object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:25:28
filename: /Users/mark/dev/R/autoGoogleAPI/googleplusDomainsv1.auto/R/plusDomains_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Acl Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

