% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Activity.actor}
\alias{Activity.actor}
\title{Activity.actor Object}
\usage{
Activity.actor(Activity.actor.clientSpecificActorInfo = NULL,
  Activity.actor.clientSpecificActorInfo.youtubeActorInfo = NULL,
  Activity.actor.image = NULL, Activity.actor.name = NULL,
  Activity.actor.verification = NULL, clientSpecificActorInfo = NULL,
  displayName = NULL, id = NULL, image = NULL, name = NULL,
  url = NULL, verification = NULL)
}
\arguments{
\item{Activity.actor.clientSpecificActorInfo}{The \link{Activity.actor.clientSpecificActorInfo} object or list of objects}

\item{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}{The \link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo} object or list of objects}

\item{Activity.actor.image}{The \link{Activity.actor.image} object or list of objects}

\item{Activity.actor.name}{The \link{Activity.actor.name} object or list of objects}

\item{Activity.actor.verification}{The \link{Activity.actor.verification} object or list of objects}

\item{clientSpecificActorInfo}{Actor info specific to particular clients}

\item{displayName}{The name of the actor, suitable for display}

\item{id}{The ID of the actor's Person resource}

\item{image}{The image representation of the actor}

\item{name}{An object representation of the individual components of name}

\item{url}{The link to the actor's Google profile}

\item{verification}{Verification status of actor}
}
\value{
Activity.actor object
}
\description{
Activity.actor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The person who performed this activity.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.previewThumbnails}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object.statusForViewer}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}}, \code{\link{Activity}},
  \code{\link{activities.insert}}
}

