% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Activity.object.actor}
\alias{Activity.object.actor}
\title{Activity.object.actor Object}
\usage{
Activity.object.actor(Activity.object.actor.clientSpecificActorInfo = NULL,
  Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo = NULL,
  Activity.object.actor.image = NULL,
  Activity.object.actor.verification = NULL, clientSpecificActorInfo = NULL,
  displayName = NULL, id = NULL, image = NULL, url = NULL,
  verification = NULL)
}
\arguments{
\item{Activity.object.actor.clientSpecificActorInfo}{The \link{Activity.object.actor.clientSpecificActorInfo} object or list of objects}

\item{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}{The \link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo} object or list of objects}

\item{Activity.object.actor.image}{The \link{Activity.object.actor.image} object or list of objects}

\item{Activity.object.actor.verification}{The \link{Activity.object.actor.verification} object or list of objects}

\item{clientSpecificActorInfo}{Actor info specific to particular clients}

\item{displayName}{The original actor's name, which is suitable for display}

\item{id}{ID of the original actor}

\item{image}{The image representation of the original actor}

\item{url}{A link to the original actor's Google profile}

\item{verification}{Verification status of actor}
}
\value{
Activity.object.actor object
}
\description{
Activity.object.actor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If this activity's object is itself another activity, such as when a person reshares an activity, this property specifies the original activity's actor.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.previewThumbnails}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object.statusForViewer}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}}, \code{\link{Activity}},
  \code{\link{activities.insert}}
}

