% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Activity.object.attachments.embed}
\alias{Activity.object.attachments.embed}
\title{Activity.object.attachments.embed Object}
\usage{
Activity.object.attachments.embed(type = NULL, url = NULL)
}
\arguments{
\item{type}{Media type of the link}

\item{url}{URL of the link}
}
\value{
Activity.object.attachments.embed object
}
\description{
Activity.object.attachments.embed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If the attachment is a video, the embeddable link.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.previewThumbnails}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object.statusForViewer}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}}, \code{\link{Activity}},
  \code{\link{activities.insert}}
}

