% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Activity.object.attachments}
\alias{Activity.object.attachments}
\title{Activity.object.attachments Object}
\usage{
Activity.object.attachments(Activity.object.attachments.embed = NULL,
  Activity.object.attachments.fullImage = NULL,
  Activity.object.attachments.image = NULL,
  Activity.object.attachments.previewThumbnails = NULL,
  Activity.object.attachments.thumbnails = NULL,
  Activity.object.attachments.thumbnails.image = NULL)
}
\arguments{
\item{Activity.object.attachments.embed}{The \link{Activity.object.attachments.embed} object or list of objects}

\item{Activity.object.attachments.fullImage}{The \link{Activity.object.attachments.fullImage} object or list of objects}

\item{Activity.object.attachments.image}{The \link{Activity.object.attachments.image} object or list of objects}

\item{Activity.object.attachments.previewThumbnails}{The \link{Activity.object.attachments.previewThumbnails} object or list of objects}

\item{Activity.object.attachments.thumbnails}{The \link{Activity.object.attachments.thumbnails} object or list of objects}

\item{Activity.object.attachments.thumbnails.image}{The \link{Activity.object.attachments.thumbnails.image} object or list of objects}
}
\value{
Activity.object.attachments object
}
\description{
Activity.object.attachments Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.previewThumbnails}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object.statusForViewer}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}}, \code{\link{Activity}},
  \code{\link{activities.insert}}
}

