% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Activity.object}
\alias{Activity.object}
\title{Activity.object Object}
\usage{
Activity.object(Activity.object.actor = NULL,
  Activity.object.actor.clientSpecificActorInfo = NULL,
  Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo = NULL,
  Activity.object.actor.image = NULL,
  Activity.object.actor.verification = NULL,
  Activity.object.attachments = NULL,
  Activity.object.attachments.embed = NULL,
  Activity.object.attachments.fullImage = NULL,
  Activity.object.attachments.image = NULL,
  Activity.object.attachments.previewThumbnails = NULL,
  Activity.object.attachments.thumbnails = NULL,
  Activity.object.attachments.thumbnails.image = NULL,
  Activity.object.plusoners = NULL, Activity.object.replies = NULL,
  Activity.object.resharers = NULL, Activity.object.statusForViewer = NULL,
  actor = NULL, attachments = NULL, content = NULL, id = NULL,
  objectType = NULL, originalContent = NULL, plusoners = NULL,
  replies = NULL, resharers = NULL, statusForViewer = NULL, url = NULL)
}
\arguments{
\item{Activity.object.actor}{The \link{Activity.object.actor} object or list of objects}

\item{Activity.object.actor.clientSpecificActorInfo}{The \link{Activity.object.actor.clientSpecificActorInfo} object or list of objects}

\item{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}{The \link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo} object or list of objects}

\item{Activity.object.actor.image}{The \link{Activity.object.actor.image} object or list of objects}

\item{Activity.object.actor.verification}{The \link{Activity.object.actor.verification} object or list of objects}

\item{Activity.object.attachments}{The \link{Activity.object.attachments} object or list of objects}

\item{Activity.object.attachments.embed}{The \link{Activity.object.attachments.embed} object or list of objects}

\item{Activity.object.attachments.fullImage}{The \link{Activity.object.attachments.fullImage} object or list of objects}

\item{Activity.object.attachments.image}{The \link{Activity.object.attachments.image} object or list of objects}

\item{Activity.object.attachments.previewThumbnails}{The \link{Activity.object.attachments.previewThumbnails} object or list of objects}

\item{Activity.object.attachments.thumbnails}{The \link{Activity.object.attachments.thumbnails} object or list of objects}

\item{Activity.object.attachments.thumbnails.image}{The \link{Activity.object.attachments.thumbnails.image} object or list of objects}

\item{Activity.object.plusoners}{The \link{Activity.object.plusoners} object or list of objects}

\item{Activity.object.replies}{The \link{Activity.object.replies} object or list of objects}

\item{Activity.object.resharers}{The \link{Activity.object.resharers} object or list of objects}

\item{Activity.object.statusForViewer}{The \link{Activity.object.statusForViewer} object or list of objects}

\item{actor}{If this activity's object is itself another activity, such as when a person reshares an activity, this property specifies the original activity's actor}

\item{attachments}{The media objects attached to this activity}

\item{content}{The HTML-formatted content, which is suitable for display}

\item{id}{The ID of the object}

\item{objectType}{The type of the object}

\item{originalContent}{The content (text) as provided by the author, which is stored without any HTML formatting}

\item{plusoners}{People who +1'd this activity}

\item{replies}{Comments in reply to this activity}

\item{resharers}{People who reshared this activity}

\item{statusForViewer}{Status of the activity as seen by the viewer}

\item{url}{The URL that points to the linked resource}
}
\value{
Activity.object object
}
\description{
Activity.object Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The object of this activity.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.previewThumbnails}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object.statusForViewer}},
  \code{\link{Activity.provider}}, \code{\link{Activity}},
  \code{\link{activities.insert}}
}
