% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Activity.object.resharers}
\alias{Activity.object.resharers}
\title{Activity.object.resharers Object}
\usage{
Activity.object.resharers(selfLink = NULL, totalItems = NULL)
}
\arguments{
\item{selfLink}{The URL for the collection of resharers}

\item{totalItems}{Total number of people who reshared this activity}
}
\value{
Activity.object.resharers object
}
\description{
Activity.object.resharers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
People who reshared this activity.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.previewThumbnails}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.statusForViewer}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}}, \code{\link{Activity}},
  \code{\link{activities.insert}}
}
