% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Activity.object.statusForViewer}
\alias{Activity.object.statusForViewer}
\title{Activity.object.statusForViewer Object}
\usage{
Activity.object.statusForViewer(canComment = NULL, canPlusone = NULL,
  canUpdate = NULL, isPlusOned = NULL, resharingDisabled = NULL)
}
\arguments{
\item{canComment}{Whether the viewer can comment on the activity}

\item{canPlusone}{Whether the viewer can +1 the activity}

\item{canUpdate}{Whether the viewer can edit or delete the activity}

\item{isPlusOned}{Whether the viewer has +1'd the activity}

\item{resharingDisabled}{Whether reshares are disabled for the activity}
}
\value{
Activity.object.statusForViewer object
}
\description{
Activity.object.statusForViewer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Status of the activity as seen by the viewer.
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.previewThumbnails}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}}, \code{\link{Activity}},
  \code{\link{activities.insert}}
}
