% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Activity}
\alias{Activity}
\title{Activity Object}
\usage{
Activity(Activity.actor = NULL,
  Activity.actor.clientSpecificActorInfo = NULL,
  Activity.actor.clientSpecificActorInfo.youtubeActorInfo = NULL,
  Activity.actor.image = NULL, Activity.actor.name = NULL,
  Activity.actor.verification = NULL, Activity.object = NULL,
  Activity.object.actor = NULL,
  Activity.object.actor.clientSpecificActorInfo = NULL,
  Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo = NULL,
  Activity.object.actor.image = NULL,
  Activity.object.actor.verification = NULL,
  Activity.object.attachments = NULL,
  Activity.object.attachments.embed = NULL,
  Activity.object.attachments.fullImage = NULL,
  Activity.object.attachments.image = NULL,
  Activity.object.attachments.previewThumbnails = NULL,
  Activity.object.attachments.thumbnails = NULL,
  Activity.object.attachments.thumbnails.image = NULL,
  Activity.object.plusoners = NULL, Activity.object.replies = NULL,
  Activity.object.resharers = NULL, Activity.object.statusForViewer = NULL,
  Activity.provider = NULL, access = NULL, actor = NULL, address = NULL,
  annotation = NULL, crosspostSource = NULL, etag = NULL,
  geocode = NULL, id = NULL, location = NULL, object = NULL,
  placeId = NULL, placeName = NULL, provider = NULL, published = NULL,
  radius = NULL, title = NULL, updated = NULL, url = NULL,
  verb = NULL)
}
\arguments{
\item{Activity.actor}{The \link{Activity.actor} object or list of objects}

\item{Activity.actor.clientSpecificActorInfo}{The \link{Activity.actor.clientSpecificActorInfo} object or list of objects}

\item{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}{The \link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo} object or list of objects}

\item{Activity.actor.image}{The \link{Activity.actor.image} object or list of objects}

\item{Activity.actor.name}{The \link{Activity.actor.name} object or list of objects}

\item{Activity.actor.verification}{The \link{Activity.actor.verification} object or list of objects}

\item{Activity.object}{The \link{Activity.object} object or list of objects}

\item{Activity.object.actor}{The \link{Activity.object.actor} object or list of objects}

\item{Activity.object.actor.clientSpecificActorInfo}{The \link{Activity.object.actor.clientSpecificActorInfo} object or list of objects}

\item{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}{The \link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo} object or list of objects}

\item{Activity.object.actor.image}{The \link{Activity.object.actor.image} object or list of objects}

\item{Activity.object.actor.verification}{The \link{Activity.object.actor.verification} object or list of objects}

\item{Activity.object.attachments}{The \link{Activity.object.attachments} object or list of objects}

\item{Activity.object.attachments.embed}{The \link{Activity.object.attachments.embed} object or list of objects}

\item{Activity.object.attachments.fullImage}{The \link{Activity.object.attachments.fullImage} object or list of objects}

\item{Activity.object.attachments.image}{The \link{Activity.object.attachments.image} object or list of objects}

\item{Activity.object.attachments.previewThumbnails}{The \link{Activity.object.attachments.previewThumbnails} object or list of objects}

\item{Activity.object.attachments.thumbnails}{The \link{Activity.object.attachments.thumbnails} object or list of objects}

\item{Activity.object.attachments.thumbnails.image}{The \link{Activity.object.attachments.thumbnails.image} object or list of objects}

\item{Activity.object.plusoners}{The \link{Activity.object.plusoners} object or list of objects}

\item{Activity.object.replies}{The \link{Activity.object.replies} object or list of objects}

\item{Activity.object.resharers}{The \link{Activity.object.resharers} object or list of objects}

\item{Activity.object.statusForViewer}{The \link{Activity.object.statusForViewer} object or list of objects}

\item{Activity.provider}{The \link{Activity.provider} object or list of objects}

\item{access}{Identifies who has access to see this activity}

\item{actor}{The person who performed this activity}

\item{address}{Street address where this activity occurred}

\item{annotation}{Additional content added by the person who shared this activity, applicable only when resharing an activity}

\item{crosspostSource}{If this activity is a crosspost from another system, this property specifies the ID of the original activity}

\item{etag}{ETag of this response for caching purposes}

\item{geocode}{Latitude and longitude where this activity occurred}

\item{id}{The ID of this activity}

\item{location}{The location where this activity occurred}

\item{object}{The object of this activity}

\item{placeId}{ID of the place where this activity occurred}

\item{placeName}{Name of the place where this activity occurred}

\item{provider}{The service provider that initially published this activity}

\item{published}{The time at which this activity was initially published}

\item{radius}{Radius, in meters, of the region where this activity occurred, centered at the latitude and longitude identified in geocode}

\item{title}{Title of this activity}

\item{updated}{The time at which this activity was last updated}

\item{url}{The link to this activity}

\item{verb}{This activity's verb, which indicates the action that was performed}
}
\value{
Activity object
}
\description{
Activity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Activity functions: \code{\link{Activity.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.actor.clientSpecificActorInfo}},
  \code{\link{Activity.actor.image}},
  \code{\link{Activity.actor.name}},
  \code{\link{Activity.actor.verification}},
  \code{\link{Activity.actor}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Activity.object.actor.clientSpecificActorInfo}},
  \code{\link{Activity.object.actor.image}},
  \code{\link{Activity.object.actor.verification}},
  \code{\link{Activity.object.actor}},
  \code{\link{Activity.object.attachments.embed}},
  \code{\link{Activity.object.attachments.fullImage}},
  \code{\link{Activity.object.attachments.image}},
  \code{\link{Activity.object.attachments.previewThumbnails}},
  \code{\link{Activity.object.attachments.thumbnails.image}},
  \code{\link{Activity.object.attachments.thumbnails}},
  \code{\link{Activity.object.attachments}},
  \code{\link{Activity.object.plusoners}},
  \code{\link{Activity.object.replies}},
  \code{\link{Activity.object.resharers}},
  \code{\link{Activity.object.statusForViewer}},
  \code{\link{Activity.object}},
  \code{\link{Activity.provider}},
  \code{\link{activities.insert}}
}

