% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{ActivityFeed}
\alias{ActivityFeed}
\title{ActivityFeed Object}
\usage{
ActivityFeed(etag = NULL, id = NULL, items = NULL, nextLink = NULL,
  nextPageToken = NULL, selfLink = NULL, title = NULL, updated = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{id}{The ID of this collection of activities}

\item{items}{The activities in this page of results}

\item{nextLink}{Link to the next page of activities}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{selfLink}{Link to this activity resource}

\item{title}{The title of this collection of activities, which is a truncated portion of the content}

\item{updated}{The time at which this collection of activities was last updated}
}
\value{
ActivityFeed object
}
\description{
ActivityFeed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

