% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Audience}
\alias{Audience}
\title{Audience Object}
\usage{
Audience(etag = NULL, item = NULL, memberCount = NULL,
  visibility = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{item}{The access control list entry}

\item{memberCount}{The number of people in this circle}

\item{visibility}{The circle members' visibility as chosen by the owner of the circle}
}
\value{
Audience object
}
\description{
Audience Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
