% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{AudiencesFeed}
\alias{AudiencesFeed}
\title{AudiencesFeed Object}
\usage{
AudiencesFeed(etag = NULL, items = NULL, nextPageToken = NULL,
  totalItems = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The audiences in this result}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{totalItems}{The total number of ACL entries}
}
\value{
AudiencesFeed object
}
\description{
AudiencesFeed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
