% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Circle.people}
\alias{Circle.people}
\title{Circle.people Object}
\usage{
Circle.people(totalItems = NULL)
}
\arguments{
\item{totalItems}{The total number of people in this circle}
}
\value{
Circle.people object
}
\description{
Circle.people Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The people in this circle.
}
\seealso{
Other Circle functions: \code{\link{Circle}},
  \code{\link{circles.insert}},
  \code{\link{circles.patch}}, \code{\link{circles.update}}
}

