% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Circle}
\alias{Circle}
\title{Circle Object}
\usage{
Circle(Circle.people = NULL, description = NULL, displayName = NULL,
  etag = NULL, id = NULL, people = NULL, selfLink = NULL)
}
\arguments{
\item{Circle.people}{The \link{Circle.people} object or list of objects}

\item{description}{The description of this circle}

\item{displayName}{The circle name}

\item{etag}{ETag of this response for caching purposes}

\item{id}{The ID of the circle}

\item{people}{The people in this circle}

\item{selfLink}{Link to this circle resource}
}
\value{
Circle object
}
\description{
Circle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Circle functions: \code{\link{Circle.people}},
  \code{\link{circles.insert}},
  \code{\link{circles.patch}}, \code{\link{circles.update}}
}
