% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{CircleFeed}
\alias{CircleFeed}
\title{CircleFeed Object}
\usage{
CircleFeed(etag = NULL, items = NULL, nextLink = NULL,
  nextPageToken = NULL, selfLink = NULL, title = NULL,
  totalItems = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The circles in this page of results}

\item{nextLink}{Link to the next page of circles}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{selfLink}{Link to this page of circles}

\item{title}{The title of this list of resources}

\item{totalItems}{The total number of circles}
}
\value{
CircleFeed object
}
\description{
CircleFeed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

