% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_functions.R
\name{circles.addPeople}
\alias{circles.addPeople}
\title{Add a person to a circle. Google+ limits certain circle operations, including the number of circle adds. Learn More.}
\usage{
circles.addPeople(circleId, email = NULL, userId = NULL)
}
\arguments{
\item{circleId}{The ID of the circle to add the person to}

\item{email}{Email of the people to add to the circle}

\item{userId}{IDs of the people to add to the circle}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/plus.circles.write
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/plus.circles.write, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/+/domains/}{Google Documentation}
}
