% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_functions.R
\name{circles.insert}
\alias{circles.insert}
\title{Create a new circle for the authenticated user.}
\usage{
circles.insert(Circle, userId)
}
\arguments{
\item{Circle}{The \link{Circle} object to pass to this method}

\item{userId}{The ID of the user to create the circle on behalf of}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/plus.circles.write
\item https://www.googleapis.com/auth/plus.login
\item https://www.googleapis.com/auth/plus.me
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/plus.circles.write, https://www.googleapis.com/auth/plus.login, https://www.googleapis.com/auth/plus.me)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/+/domains/}{Google Documentation}

Other Circle functions: \code{\link{Circle.people}},
  \code{\link{Circle}}, \code{\link{circles.patch}},
  \code{\link{circles.update}}
}
