% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Comment.object}
\alias{Comment.object}
\title{Comment.object Object}
\usage{
Comment.object(content = NULL, objectType = NULL, originalContent = NULL)
}
\arguments{
\item{content}{The HTML-formatted content, suitable for display}

\item{objectType}{The object type of this comment}

\item{originalContent}{The content (text) as provided by the author, stored without any HTML formatting}
}
\value{
Comment.object object
}
\description{
Comment.object Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The object of this comment.
}
\seealso{
Other Comment functions: \code{\link{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Comment.actor.clientSpecificActorInfo}},
  \code{\link{Comment.actor.image}},
  \code{\link{Comment.actor.verification}},
  \code{\link{Comment.actor}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.plusoners}}, \code{\link{Comment}},
  \code{\link{comments.insert}}
}
