% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Comment.plusoners}
\alias{Comment.plusoners}
\title{Comment.plusoners Object}
\usage{
Comment.plusoners(totalItems = NULL)
}
\arguments{
\item{totalItems}{Total number of people who +1'd this comment}
}
\value{
Comment.plusoners object
}
\description{
Comment.plusoners Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
People who +1'd this comment.
}
\seealso{
Other Comment functions: \code{\link{Comment.actor.clientSpecificActorInfo.youtubeActorInfo}},
  \code{\link{Comment.actor.clientSpecificActorInfo}},
  \code{\link{Comment.actor.image}},
  \code{\link{Comment.actor.verification}},
  \code{\link{Comment.actor}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.object}}, \code{\link{Comment}},
  \code{\link{comments.insert}}
}

