% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{CommentFeed}
\alias{CommentFeed}
\title{CommentFeed Object}
\usage{
CommentFeed(etag = NULL, id = NULL, items = NULL, nextLink = NULL,
  nextPageToken = NULL, title = NULL, updated = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{id}{The ID of this collection of comments}

\item{items}{The comments in this page of results}

\item{nextLink}{Link to the next page of activities}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{title}{The title of this collection of comments}

\item{updated}{The time at which this collection of comments was last updated}
}
\value{
CommentFeed object
}
\description{
CommentFeed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
