% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_functions.R
\name{comments.list}
\alias{comments.list}
\title{List all of the comments for an activity.}
\usage{
comments.list(activityId, maxResults = NULL, pageToken = NULL,
  sortOrder = NULL)
}
\arguments{
\item{activityId}{The ID of the activity to get comments for}

\item{maxResults}{The maximum number of comments to include in the response, which is used for paging}

\item{pageToken}{The continuation token, which is used to page through large result sets}

\item{sortOrder}{The order in which to sort the list of comments}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/plus.login
\item https://www.googleapis.com/auth/plus.stream.read
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/plus.login, https://www.googleapis.com/auth/plus.stream.read)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/+/domains/}{Google Documentation}
}
