% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Media.author.image}
\alias{Media.author.image}
\title{Media.author.image Object}
\usage{
Media.author.image(url = NULL)
}
\arguments{
\item{url}{The URL of the author's profile photo}
}
\value{
Media.author.image object
}
\description{
Media.author.image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The author's Google profile image.
}
\seealso{
Other Media functions: \code{\link{Media.author}},
  \code{\link{Media.exif}}, \code{\link{Media}},
  \code{\link{media.insert}}
}

