% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Media.author}
\alias{Media.author}
\title{Media.author Object}
\usage{
Media.author(Media.author.image = NULL, displayName = NULL, id = NULL,
  image = NULL, url = NULL)
}
\arguments{
\item{Media.author.image}{The \link{Media.author.image} object or list of objects}

\item{displayName}{The author's name}

\item{id}{ID of the author}

\item{image}{The author's Google profile image}

\item{url}{A link to the author's Google profile}
}
\value{
Media.author object
}
\description{
Media.author Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The person who uploaded this media.
}
\seealso{
Other Media functions: \code{\link{Media.author.image}},
  \code{\link{Media.exif}}, \code{\link{Media}},
  \code{\link{media.insert}}
}
