% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Media.exif}
\alias{Media.exif}
\title{Media.exif Object}
\usage{
Media.exif(time = NULL)
}
\arguments{
\item{time}{The time the media was captured}
}
\value{
Media.exif object
}
\description{
Media.exif Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Exif information of the media item.
}
\seealso{
Other Media functions: \code{\link{Media.author.image}},
  \code{\link{Media.author}}, \code{\link{Media}},
  \code{\link{media.insert}}
}
