% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Media}
\alias{Media}
\title{Media Object}
\usage{
Media(Media.author = NULL, Media.author.image = NULL, Media.exif = NULL,
  author = NULL, displayName = NULL, etag = NULL, exif = NULL,
  height = NULL, id = NULL, mediaCreatedTime = NULL, mediaUrl = NULL,
  published = NULL, sizeBytes = NULL, streams = NULL, summary = NULL,
  updated = NULL, url = NULL, videoDuration = NULL, videoStatus = NULL,
  width = NULL)
}
\arguments{
\item{Media.author}{The \link{Media.author} object or list of objects}

\item{Media.author.image}{The \link{Media.author.image} object or list of objects}

\item{Media.exif}{The \link{Media.exif} object or list of objects}

\item{author}{The person who uploaded this media}

\item{displayName}{The display name for this media}

\item{etag}{ETag of this response for caching purposes}

\item{exif}{Exif information of the media item}

\item{height}{The height in pixels of the original image}

\item{id}{ID of this media, which is generated by the API}

\item{mediaCreatedTime}{The time at which this media was originally created in UTC}

\item{mediaUrl}{The URL of this photo or video's still image}

\item{published}{The time at which this media was uploaded}

\item{sizeBytes}{The size in bytes of this video}

\item{streams}{The list of video streams for this video}

\item{summary}{A description, or caption, for this media}

\item{updated}{The time at which this media was last updated}

\item{url}{The URL for the page that hosts this media}

\item{videoDuration}{The duration in milliseconds of this video}

\item{videoStatus}{The encoding status of this video}

\item{width}{The width in pixels of the original image}
}
\value{
Media object
}
\description{
Media Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Media functions: \code{\link{Media.author.image}},
  \code{\link{Media.author}}, \code{\link{Media.exif}},
  \code{\link{media.insert}}
}
