% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_functions.R
\name{people.listByCircle}
\alias{people.listByCircle}
\title{List all of the people who are members of a circle.}
\usage{
people.listByCircle(circleId, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{circleId}{The ID of the circle to get the members of}

\item{maxResults}{The maximum number of people to include in the response, which is used for paging}

\item{pageToken}{The continuation token, which is used to page through large result sets}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/plus.circles.read
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/plus.circles.read, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/+/domains/}{Google Documentation}
}
