% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{PeopleFeed}
\alias{PeopleFeed}
\title{PeopleFeed Object}
\usage{
PeopleFeed(etag = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL, title = NULL, totalItems = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The people in this page of results}

\item{nextPageToken}{The continuation token, which is used to page through large result sets}

\item{selfLink}{Link to this resource}

\item{title}{The title of this collection of people}

\item{totalItems}{The total number of people available in this list}
}
\value{
PeopleFeed object
}
\description{
PeopleFeed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

