% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Place}
\alias{Place}
\title{Place Object}
\usage{
Place(Place.address = NULL, Place.position = NULL, address = NULL,
  displayName = NULL, id = NULL, position = NULL)
}
\arguments{
\item{Place.address}{The \link{Place.address} object or list of objects}

\item{Place.position}{The \link{Place.position} object or list of objects}

\item{address}{The physical address of the place}

\item{displayName}{The display name of the place}

\item{id}{The id of the place}

\item{position}{The position of the place}
}
\value{
Place object
}
\description{
Place Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Place functions: \code{\link{Place.address}},
  \code{\link{Place.position}}
}
