% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_functions.R
\docType{package}
\name{plusDomains_googleAuthR}
\alias{plusDomains_googleAuthR}
\alias{plusDomains_googleAuthR-package}
\title{Google+ Domains API
Builds on top of the Google+ platform for Google Apps Domains.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:25:27
filename: /Users/mark/dev/R/autoGoogleAPI/googleplusDomainsv1.auto/R/plusDomains_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/plus.circles.read
\item https://www.googleapis.com/auth/plus.circles.write
\item https://www.googleapis.com/auth/plus.login
\item https://www.googleapis.com/auth/plus.me
\item https://www.googleapis.com/auth/plus.media.upload
\item https://www.googleapis.com/auth/plus.profiles.read
\item https://www.googleapis.com/auth/plus.stream.read
\item https://www.googleapis.com/auth/plus.stream.write
\item https://www.googleapis.com/auth/userinfo.email
\item https://www.googleapis.com/auth/userinfo.profile
}
}

