% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusDomains_objects.R
\name{Videostream}
\alias{Videostream}
\title{Videostream Object}
\usage{
Videostream(height = NULL, type = NULL, url = NULL, width = NULL)
}
\arguments{
\item{height}{The height, in pixels, of the video resource}

\item{type}{MIME type of the video stream}

\item{url}{URL of the video stream}

\item{width}{The width, in pixels, of the video resource}
}
\value{
Videostream object
}
\description{
Videostream Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
